Fixed an Error and added UniverseHost
All businesses and legal entities in the Netherlands are required to register in the Commercial Register. According to the law, you operate a business if you independently supply goods or services to other parties in order to make a profit. UniverseHost is in violation of this law as evidence shows here: http://prntscr.com/jpwgct